create a checkout component
